Title: 숫자 출력하기, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 250221/숫자 출력하기/print-one-number.cpp
@@ -3,6 +3,6 @@ using namespace std;
 
 int main() {
     // Please write your code here.
-    cout << 3;
+    cout << (int)3;
     return 0;
 }
